requirements.txt
dateutils
azure-storage-blob==12.19.1
azure-identity==1.16.0
azure-datalake-store==0.0.53
millify==0.1.1
numpy==1.26.4
openai==1.34.0
pandas==1.5.3
plotly==5.22.0
python-dotenv==1.0.1
pandas==1.5.3
plotly==5.22.0
python-dotenv==1.0.1
streamlit==1.35.0
streamlit-card
streamlit-extras
seaborn==0.13.2
duckdb==0.10.3
pygwalker==0.4.8.9
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.11
langchain-experimental==0.0.62
langchain-openai==0.1.14
openai==1.34.0
wordcloud==1.9.3
tabulate==0.9.0
polars==0.20.23
smart-open==7.0.5
fsspec==2024.3.1